requirements.txt
discord.py==0.16.12
tensorflow==1.14.0